[ui] Refactor OS warning UI into separate file
Create a new window-status-displaying.js and move the logic that
displays the "ChromeOS only" warning into it.

This is a pure refactoring change. It's a preparation for implementing
crash warning UI as tracked by #926.
